mini project
This mini-project is based on Graph Theory and aims to study the properties and performance of graphs using C programming and MATLAB.

Graphs are mathematical structures used to model relationships between objects. In this project, we will randomly generate graphs, analyze their properties, and evaluate the computational time needed for different graph sizes.